requirements.txt
pytest-playwright==0.4.4
playwright==1.42.0
allure-pytest==2.13.3
pytest==7.4.0
selenium==4.18.1
allure-python-commons==2.13.2